Set up the keyboard when constructing the instrument
