+++ src/instrument.cpp
@@ -1,4 +1,6 @@
 #include "instrument.h"
+#include "constants.h"
+#include "helpers.h"
 
 namespace blga {
 
@@ -9,18 +11,25 @@ instrument::instrument(
     : d_envelope(envelope)
     , d_oscillator(oscillator)
 {
+    // Create the keyboard
+    for (auto [index, key] : blga::enumerate(blga::keyboard)) {
+        d_notes[key] = {
+            .frequency = blga::note_frequency(3, key_name{index}),
+            .toggle_time = -1.0,
+            .active = false
+        };
+    }
 }
 
-auto instrument::note_on(char note, double dt, double frequency) -> void
+auto instrument::note_on(char note, double dt) -> void
 {
-    d_notes[note].age = dt;
+    d_notes[note].toggle_time = dt;
     d_notes[note].active = true;
-    d_notes[note].frequency = frequency;
 }
 
 auto instrument::note_off(char note, double dt) -> void
 {
-    d_notes[note].age = dt;
+    d_notes[note].toggle_time = dt;
     d_notes[note].active = false;
 }
 
@@ -29,7 +38,7 @@ auto instrument::amplitude(double dt) -> double
     double amp = 0.0;
     for (const auto& [key, note] : d_notes) {
         amp += (
-            d_envelope.amplitude(dt, note.age, note.active) *
+            d_envelope.amplitude(dt, note.toggle_time, note.active) *
             d_oscillator(note.frequency, dt)
         );
     }
